+ Make `ErrorMessage` cross-platform + Make `Mask` cross-platform + Make `LoadingIndicator` cross-platform
