+++ cmp/error/ErrorBoundary.ts
@@ -5,8 +5,7 @@
  * Copyright © 2024 Extremely Heavy Industries Inc.
  */
 import {elementFactory, hoistCmp, uses, XH} from '@xh/hoist/core';
-import {errorMessage as desktopErrorMessage} from '@xh/hoist/dynamics/desktop';
-import {errorMessage as mobileErrorMessage} from '@xh/hoist/dynamics/mobile';
+import {errorMessage} from '@xh/hoist/cmp/error';
 import {Component, ReactNode} from 'react';
 import {ErrorBoundaryModel} from './ErrorBoundaryModel';
 
@@ -34,8 +33,7 @@ export const [ErrorBoundary, errorBoundary] = hoistCmp.withFactory<ErrorBoundary
         if (!error) return reactErrorBoundary({model, children});
         if (errorRenderer) return errorRenderer(error);
 
-        const cmp = XH.isMobileApp ? mobileErrorMessage : desktopErrorMessage;
-        return cmp({
+        return errorMessage({
             error,
             title: 'Unexpected error while rendering this component',
             actionFn: () => model.clear(),
