+++ cmp/error/index.ts
@@ -1,2 +1,3 @@
 export * from './ErrorBoundary';
 export * from './ErrorBoundaryModel';
+export * from './ErrorMessage';
