+++ cmp/loadingindicator/index.ts
