+++ cmp/mask/index.ts
