+++ desktop/cmp/mask/impl/Mask.scss
