--- mobile/cmp/loadingindicator/index.ts
@@ -1,7 +1,0 @@
-/*
- * This file belongs to Hoist, an application development toolkit
- * developed by Extremely Heavy Industries (www.xh.io | info@xh.io)
- *
- * Copyright © 2024 Extremely Heavy Industries Inc.
- */
-export * from './LoadingIndicator';
