+++ mobile/cmp/mask/impl/Mask.scss
